More fixes for resolving warnings from the Intel compiler.  These fixes resolve a subset of the remark #2259 warnings:    remark #2259: non-pointer conversion from "double" to "int" may lose significant bits
There are additional remark #2259 warnings that are yet to be addressed.

+++ src/basic/vx_util/data_plane_util.cc
@@ -256,7 +256,7 @@ void fractional_coverage(const DataPlane &dp, DataPlane &frac_dp,
                yy = y_ur;
 
                // Decrement counts for data to be replaced
-               k = box_na[n];
+               k = nint(box_na[n]);
                if(!is_bad_data(k)) {
                   count_vld -= 1;
                   count_thr -= k;
